# Populate the Cshift_table in the GPU