@paboyle 2023-03-24
THANKS !!!
This looks like a great speed up.